Refactoring: GridCreatingConditionTriangle::AddDivisionLineCommand implementation is separated from gridcreatingconditiontriangle.cpp.
